Fix overflow issue with large matrices (fixes #4595)

+++ src/shogun/statistical_testing/internals/mmd/ComputeMMD.h
@@ -108,13 +108,14 @@ struct ComputeMMD
 		const BlockXt& b_xy=map.block(m_n_x, 0, m_n_y, m_n_x);
 
 		terms_t terms;
-		terms.diag[0]=b_x.diagonal().sum();
-		terms.diag[1]=b_y.diagonal().sum();
-		terms.diag[2]=b_xy.diagonal().sum();
+		// cast to 64 bit to avoid overflows
+		terms.diag[0]=b_x.diagonal(). template cast<float64_t>().sum();
+		terms.diag[1]=b_y.diagonal(). template cast<float64_t>().sum();
+		terms.diag[2]=b_xy.diagonal(). template cast<float64_t>().sum();
 
-		terms.term[0]=(b_x.sum()-terms.diag[0])/2+terms.diag[0];
-		terms.term[1]=(b_y.sum()-terms.diag[1])/2+terms.diag[1];
-		terms.term[2]=b_xy.sum();
+		terms.term[0]=(b_x. template cast<float64_t>().sum()-terms.diag[0])/2+terms.diag[0];
+		terms.term[1]=(b_y. template cast<float64_t>().sum()-terms.diag[1])/2+terms.diag[1];
+		terms.term[2]=b_xy. template cast<float64_t>().sum();
 
 		return compute(terms);
 	}
